Remove experimental GestureOptions.pinchBehavior
* This change is in preparation for a subsequent change to only send
  .cameraChanged events once per update in PinchGestureHandler.
