+++ Sources/MapboxMaps/Gestures/GestureHandlers/PinchGestureHandler.swift
@@ -2,86 +2,159 @@ import UIKit
 
 internal protocol PinchGestureHandlerProtocol: GestureHandler {
     var rotateEnabled: Bool { get set }
-    var behavior: PinchGestureBehavior { get set }
-}
-
-internal protocol PinchGestureHandlerImpl: AnyObject {
-    func handleGesture(_ gestureRecognizer: UIPinchGestureRecognizer, state: UIGestureRecognizer.State)
 }
 
 /// `PinchGestureHandler` updates the map camera in response to a 2-touch
 /// gesture that may consist of translation, scaling, and rotation
 internal final class PinchGestureHandler: GestureHandler, PinchGestureHandlerProtocol {
     /// Whether pinch gesture can rotate map or not
-    internal var rotateEnabled: Bool = true {
-        didSet {
-            impl1.rotateEnabled = rotateEnabled
-            impl2.rotateEnabled = rotateEnabled
-        }
-    }
+    internal var rotateEnabled: Bool = true
 
-    internal var behavior: PinchGestureBehavior = .tracksTouchLocationsWhenPanningAfterZoomChange
+    /// The midpoint of the touches in the gesture's view when the gesture began
+    private var initialPinchMidpoint: CGPoint?
 
-    private var initialBehavior: PinchGestureBehavior?
+    /// The angle from touch location 0 to touch location 1 when the gesture began or unpaused
+    private var initialPinchAngle: CGFloat?
 
-    private let impl1: PinchGestureHandlerImpl1
+    /// The camera center when the gesture began or unpaused
+    private var initialCenter: CLLocationCoordinate2D?
 
-    private let impl2: PinchGestureHandlerImpl2
+    /// The camera zoom when the gesture began
+    private var initialZoom: CGFloat?
+
+    /// The camera bearing when the gesture began or unpaused
+    private var initialBearing: CLLocationDirection?
+
+    /// The rotateEnabled setting when the gesture began
+    private var initialRotateEnabled: Bool?
+
+    private var gestureBegan = false
+
+    private let mapboxMap: MapboxMapProtocol
 
     /// Initialize the handler which creates the panGestureRecognizer and adds to the view
     internal init(gestureRecognizer: UIPinchGestureRecognizer,
                   mapboxMap: MapboxMapProtocol) {
-        self.impl1 = PinchGestureHandlerImpl1(mapboxMap: mapboxMap)
-        self.impl2 = PinchGestureHandlerImpl2(mapboxMap: mapboxMap)
+        self.mapboxMap = mapboxMap
         super.init(gestureRecognizer: gestureRecognizer)
         gestureRecognizer.addTarget(self, action: #selector(handleGesture(_:)))
-        impl1.delegate = self
-        impl2.delegate = self
     }
 
+    // swiftlint:disable:next cyclomatic_complexity
     @objc private func handleGesture(_ gestureRecognizer: UIPinchGestureRecognizer) {
-        let effectiveBehavior: PinchGestureBehavior?
+        guard let view = gestureRecognizer.view else {
+            return
+        }
+        let pinchMidpoint = gestureRecognizer.location(in: view)
 
-        let state = gestureRecognizer.state
-
-        switch state {
+        switch gestureRecognizer.state {
         case .began:
-            effectiveBehavior = behavior
-            initialBehavior = behavior
+            guard gestureRecognizer.numberOfTouches == 2 else {
+                return
+            }
+            initialPinchMidpoint = pinchMidpoint
+            initialPinchAngle = pinchAngle(with: gestureRecognizer)
+            initialCenter = mapboxMap.cameraState.center
+            initialZoom = mapboxMap.cameraState.zoom
+            initialBearing = mapboxMap.cameraState.bearing
+            initialRotateEnabled = rotateEnabled
+            gestureBegan = true
+            delegate?.gestureBegan(for: .pinch)
         case .changed:
-            effectiveBehavior = initialBehavior
-        case .cancelled, .ended:
-            effectiveBehavior = initialBehavior
-            initialBehavior = nil
+            // UIPinchGestureRecognizer sends a .changed event when the number
+            // of touches decreases from 2 to 1. If this happens, we pause our
+            // gesture handling.
+            //
+            // if a second touch goes down again before the gesture ends, we
+            // resume and re-capture the initial state (except for zoom since
+            // UIPinchGestureRecognizer provides continuity of scale values)
+            guard gestureRecognizer.numberOfTouches == 2 else {
+                initialPinchMidpoint = nil
+                initialPinchAngle = nil
+                initialCenter = nil
+                initialBearing = nil
+                return
+            }
+            guard let initialZoom = initialZoom,
+                  let initialRotateEnabled = initialRotateEnabled else {
+                return
+            }
+            // Using explicit self to help older versions of Xcode (pre-12.5) figure
+            // out the scope of these variables. https://bugs.swift.org/browse/SR-8669
+            let pinchAngle = self.pinchAngle(with: gestureRecognizer)
+            guard let initialPinchMidpoint = initialPinchMidpoint,
+                  let initialPinchAngle = initialPinchAngle,
+                  let initialCenter = initialCenter,
+                  let initialBearing = initialBearing else {
+                self.initialPinchMidpoint = pinchMidpoint
+                self.initialPinchAngle = pinchAngle
+                self.initialCenter = mapboxMap.cameraState.center
+                self.initialBearing = mapboxMap.cameraState.bearing
+                return
+            }
+
+            let zoomIncrement = log2(gestureRecognizer.scale)
+            var cameraOptions = CameraOptions()
+            cameraOptions.center = initialCenter
+            cameraOptions.zoom = initialZoom
+            cameraOptions.bearing = initialRotateEnabled ? initialBearing : nil
+
+            mapboxMap.setCamera(to: cameraOptions)
+
+            mapboxMap.dragStart(for: initialPinchMidpoint)
+            let dragOptions = mapboxMap.dragCameraOptions(
+                from: initialPinchMidpoint,
+                to: pinchMidpoint)
+            mapboxMap.setCamera(to: dragOptions)
+            mapboxMap.dragEnd()
+            // the two angles will always be in the range [0, 2pi)
+            // so the resulting rotation will be in the range (-2pi, 2pi)
+            var rotation = pinchAngle - initialPinchAngle
+            // if the rotation is negative, add 2pi so that the final
+            // result is in the range [0, 2pi)
+            if rotation < 0 {
+                rotation += 2 * .pi
+            }
+            // convert from radians to degrees and flip the sign since
+            // the iOS coordinate system is flipped relative to the
+            // coordinate system used for bearing in the map.
+            let rotationInDegrees = Double(rotation * 180.0 / .pi * -1)
+
+            mapboxMap.setCamera(
+                to: CameraOptions(
+                    anchor: pinchMidpoint,
+                    zoom: initialZoom + zoomIncrement,
+                    bearing: initialRotateEnabled ? (initialBearing + rotationInDegrees) : nil))
+        case .ended, .cancelled:
+            initialPinchMidpoint = nil
+            initialPinchAngle = nil
+            initialCenter = nil
+            initialZoom = nil
+            initialBearing = nil
+            initialRotateEnabled = nil
+            if gestureBegan {
+                delegate?.gestureEnded(for: .pinch, willAnimate: false)
+            }
+            gestureBegan = false
         default:
-            effectiveBehavior = nil
+            break
         }
+    }
 
-        let impl: PinchGestureHandlerImpl?
-
-        switch effectiveBehavior {
-        case .tracksTouchLocationsWhenPanningAfterZoomChange:
-            impl = impl1
-        case .doesNotResetCameraAtEachFrame:
-            impl = impl2
-        default:
-            impl = nil
+    /// Returns the angle in radians in the range [0, 2pi)
+    private func angleOfLine(from point0: CGPoint, to point1: CGPoint) -> CGFloat {
+        var angle = atan2(point1.y - point0.y, point1.x - point0.x)
+        if angle < 0 {
+            angle += 2 * .pi
         }
-
-        impl?.handleGesture(gestureRecognizer, state: state)
-    }
-}
-
-extension PinchGestureHandler: GestureHandlerDelegate {
-    func gestureBegan(for gestureType: GestureType) {
-        delegate?.gestureBegan(for: gestureType)
+        return angle
     }
 
-    func gestureEnded(for gestureType: GestureType, willAnimate: Bool) {
-        delegate?.gestureEnded(for: gestureType, willAnimate: willAnimate)
-    }
-
-    func animationEnded(for gestureType: GestureType) {
-        delegate?.animationEnded(for: gestureType)
+    private func pinchAngle(with gestureRecognizer: UIPinchGestureRecognizer) -> CGFloat {
+        // we guard for this at the call site
+        let view = gestureRecognizer.view!
+        let pinchPoint0 = gestureRecognizer.location(ofTouch: 0, in: view)
+        let pinchPoint1 = gestureRecognizer.location(ofTouch: 1, in: view)
+        return angleOfLine(from: pinchPoint0, to: pinchPoint1)
     }
 }
