+++ Tests/MapboxMapsTests/Gestures/GestureHandlers/Mocks/MockPinchGestureHandler.swift
@@ -1,7 +1,5 @@
-@_spi(Experimental) @testable import MapboxMaps
+@testable import MapboxMaps
 
 final class MockPinchGestureHandler: GestureHandler, PinchGestureHandlerProtocol {
     var rotateEnabled: Bool = true
-
-    var behavior: PinchGestureBehavior = .tracksTouchLocationsWhenPanningAfterZoomChange
 }
